Optimize btslogsorter implementation

+++ AprgAlgorithm/AprgAlgorithm/src/LargeSorter/DataBlockCache.hpp
@@ -26,7 +26,8 @@ public:
     using BlockCacheContainer = std::deque<BlockCacheEntry>;
 
     void addBlock(unsigned const int blockId, BlockInformation const& iterator)
-    {        if(m_blocksInformationCache.empty() || m_blocksInformationCache.front().m_blockId != blockId)
+    {
+        if(m_blocksInformationCache.empty() || m_blocksInformationCache.front().m_blockId != blockId)
         {
             deleteBlock(blockId);
             m_blocksInformationCache.emplace_front(blockId, iterator);
@@ -37,7 +38,8 @@ public:
         m_blocksInformationCache.erase(std::remove_if(m_blocksInformationCache.begin(), m_blocksInformationCache.end(), [blockId](BlockCacheEntry const& blockInformation)
         {
             return blockId == blockInformation.m_blockId;
-        }), m_blocksInformationCache.end());    }
+        }), m_blocksInformationCache.end());
+    }
     BlockInformation popTheEarliestAddedBlock()
     {
         if(!m_blocksInformationCache.empty())
@@ -51,7 +53,8 @@ public:
     BlockCacheContainer & getContainerReference()
     {
         return m_blocksInformationCache;
-    }    void clear()
+    }
+    void clear()
     {
         m_blocksInformationCache.clear();
     }
